+++ DemoParticles/Src/Content/ParticleEmitterPoint.cpp
@@ -47,7 +47,7 @@ namespace DemoParticles
         //false to reset if the previous render emit particles
         m_needEmit = false;
 
-        m_lastEmitTime -= timer.GetElapsedSeconds();
+        m_lastEmitTime -= (float)timer.GetElapsedSeconds();
         if (m_lastEmitTime <= 0.0)
         {
             m_lastEmitTime = m_emitDelay;
